+++ src/structures/VoiceState.js
@@ -83,7 +83,9 @@ class VoiceState extends Base {
       this.sessionId ??= null;
     }
 
-    if ('self_stream' in data) {
+    // The self_stream is property is omitted if false, check for another property
+    // here to avoid incorrectly clearing this when partial data is specified
+    if ('self_video' in data) {
       /**
        * Whether this member is streaming using "Screen Share"
        * @type {boolean}